fix: overhaul both image and webcam files and new files
fix: imgviz syntax error fix
feat: create const.py for store all constant variables for both files
feat: create preprocessing.py file for common functions
