+++ const.py
@@ -1,0 +1,6 @@
+LOAD_MODEL_DEPLOY_PATH = "./model_ear/best_model_ear_v1_43.pth"
+ENCODER = 'resnet18'
+ENCODER_WEIGHTS = 'imagenet'
+CLASSES = ['ear']
+ACTIVATION = 'sigmoid'
+DEVICE = "cpu"
